Card style search results
Mockup of the design of how the results of a search will be presented. They are shown in a card style with their most basic information, as well as a link to their actual product page. I don't know if said product page will be needed, could just embed a buy button on the result